+++ rest.osgi/src/main/java/com/ericsson/bss/cassandra/ecchronos/rest/osgi/RepairManagementRESTComponent.java
@@ -24,13 +24,10 @@ import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.component.annotations.ReferenceCardinality;
 import org.osgi.service.component.annotations.ReferencePolicy;
 
-import javax.ws.rs.Path;
-
 /**
  * OSGi component wrapping {@link RepairManagementREST} bound with OSGi services.
  */
 @Component
-@Path("/repair-management/v1")
 public class RepairManagementRESTComponent implements RepairManagementREST
 {
     @Reference (service = RepairScheduler.class, cardinality = ReferenceCardinality.MANDATORY, policy = ReferencePolicy.STATIC)